MSI-1840: refactored GetSourceSelectionResultFromInvoice in order to avoid static tests falling
